add things

+++ db-core/src/main/java/dev/davidsilva/blog/dbcore/post/PostDao.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PostDao extends JpaRepository<Post, Long>, JpaSpecificationExecutor<Post> {
+public interface PostDao extends JpaRepository<Post, Integer>, JpaSpecificationExecutor<Post> {
 }
